added support for API polling
